# GGRC-5854 Make a possibility to stop the export