--- a/src/ggrc/models/exceptions.py
+++ b/src/ggrc/models/exceptions.py
@@ -62,3 +62,11 @@ class ValidationError(ValueError):
 
 class StatusValidationError(ValidationError):
   pass
+
+
+class ReservedNameError(ValueError):
+  pass
+
+
+class ExportStoppedException(RuntimeError):
+  pass