--- a/src/ggrc/views/converters.py
+++ b/src/ggrc/views/converters.py
@@ -22,27 +22,36 @@ from googleapiclient import errors
 
 import flask
 from flask import current_app
-from flask import request
 from flask import json
 from flask import render_template
+from flask import request
 from werkzeug import exceptions as wzg_exceptions
 
-from ggrc.models.exceptions import ExportStoppedException
-from ggrc import db, utils
+
+from ggrc import db
+from ggrc import settings
+from ggrc import utils
 from ggrc.app import app
 from ggrc.cloud_api import task_queue
 from ggrc.converters import get_exportables
-from ggrc.converters.base import ImportConverter, ExportConverter
-from ggrc.converters.import_helper import count_objects, \
-    read_csv_file, get_export_filename, get_object_column_definitions
+from ggrc.converters.base import ExportConverter
+from ggrc.converters.base import ImportConverter
+from ggrc.converters.import_helper import count_objects
+from ggrc.converters.import_helper import get_export_filename
+from ggrc.converters.import_helper import get_object_column_definitions
+from ggrc.converters.import_helper import read_csv_file
 from ggrc.gdrive import file_actions as fa
-from ggrc.models import import_export, background_task, all_models
+from ggrc.login import get_current_user
+from ggrc.login import login_required
+from ggrc.models import all_models
+from ggrc.models import background_task
+from ggrc.models import import_export
+from ggrc.models.exceptions import ExportStoppedException
 from ggrc.notifications import job_emails
 from ggrc.query.exceptions import BadQueryException
 from ggrc.query.builder import QueryHelper
-from ggrc.login import login_required, get_current_user
-from ggrc import settings
-from ggrc.utils import benchmark, errors as app_errors
+from ggrc.utils import benchmark
+from ggrc.utils import errors as app_errors
 
 
 EXPORTABLES_MAP = {exportable.__name__: exportable for exportable