<!-- If your PR depends on other tickets or PRs, you can list them here
# Dependencies

This PR is `on hold` until the dependencies are merged:

- [ ] GGRC-1234
- [ ] #1234
-->

# Issue description
When you try to kill export using kill button it haven't stoped and will send email notification when it would be finished.

# Steps to test the changes
Start long export. Kill it. Look in Task queue that export job have stopped.
# Solution description

I've added check to export handler to check if export was stopped and interrupt process.
Also I've refactored our base.py and converters.py files. Please look through commit messages.

# Notes
## Performance test for solution with 
### memcache solution 

(1500 programs exported on perf instance. time in seconds.):
dev branch: 84.3, 63.8, 55.4, 64.5, 66.8
branch with changes: 83.4, 69.9, 65.5, 68.1, 62.3

average: 
- before 66.96
- after 69.84
- 4.3% degradation

average w/o corner values:
- before 65.05
- after 67.83
- 4.27% degradation 

### results for previous solution (db requests)
We've tested best solution we've found and it causes performance degradation approx. 4.5%. We've measured it for 995 assessments export w/o custom attributes.

# Sanity checklist

- [x] I have clicked through the app to make sure my changes work and not break the app.
- [x] I have applied the correct milestone and labels.
- [x] My changes fix the issue described in the description (and do nothing else). 🤞
- [ ] My changes are covered by tests.
- [ ] My changes follow our [performance guidelines](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/performance.rst).
- [x] My changes follow our [js](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/javascript.rst) and/or [python](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/python.rst) guidelines.
- [x] My commits follow our [commit guidelines](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/git/how_to_write_a_commit_message.rst).

<!-- If your PR includes a migration include the additional checklist items
# Migration checklist
- [ ] Migration passes all checks from our [PR review guidelines](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/git/reviewing_pull_requests.rst#reviewing-a-pr-containing-database-migration-scripts)
- [ ] Upon merging, add 'check migration chain' and 'kokoro:run' labels to all open PRs with a label 'migration'
-->

# PR Review checklist

- [x] The changes fix the issue and don't cause any apparent regressions.
- [x] Labels and milestone are correctly set.
- [x] The solution description matches the changes in the code.
- [x] There is no apparent way to improve the performance & design of the new code.
- [x] The pull request is opened against the correct base branch.
- [x] Upon merging, the Jira ticket's fixversion is correctly set and the ticket is moved to "QA - In Progress".

<!-- If your code is not finished yet can include a TODO check list
# TODO

- [ ] First item on the TODO
-->