@mashabarashko 2019-03-06
@belovmd Here are prefomance results:

Before fix: 131.81s 126.64s 135.67s 
After fix: 177.27s 179.78s 169.80s

As you can see, performance got slightly worse. I think we need confirm if it's acceptable and improve fix if it's not.