@belovmd 2019-03-21
Performance test (1500 programs exported on perf instance. time in seconds.):
dev branch: 84.3, 63.8, 55.4, 64.5, 66.8
branch with changes: 83.4, 69.9, 65.5, 68.1, 62.3

average: 
- before 66.96
- after 69.84
- 4.3% degradation

average w/o corner values:
- before 65.05
- after 67.83
- 4.27% degradation