@zidarsk8 2019-03-21
@belovmd Please keep in mind that having 5 examples is really not enough to say if this is really a degradation or not.